refactored existing gradle tests
