+++ tests/conftest.py
@@ -90,6 +90,7 @@ def first_app_config():
         version="0.0.1",
         description="The first simple app",
         sources=["src/first"],
+        packaging_format="aab",
     )
